+++ compat_test.go
@@ -1,0 +1,10 @@
+package rest
+
+import "testing"
+
+func TestCompat(t *testing.T) {
+	_ = &Error{Title: "foo"}
+	_ = &Client{}
+	d := NewClient("http://base", "user", "pass")
+	d.Client.Transport = DefaultTransport
+}