+++ rest.go
@@ -8,6 +8,20 @@ import (
 	"sync"
 
 	log "github.com/inconshreveable/log15"
+	"github.com/kevinburke/rest/restclient"
+	"github.com/kevinburke/rest/resterror"
+)
+
+type (
+	// Backwards compatibility
+	Error     = resterror.Error
+	Client    = restclient.Client
+	Transport = restclient.Transport
+)
+
+var (
+	NewClient        = restclient.New
+	DefaultTransport = restclient.DefaultTransport
 )
 
 const jsonContentType = "application/json; charset=utf-8"
@@ -15,43 +29,10 @@ const jsonContentType = "application/json; charset=utf-8"
 // Logger logs information about incoming requests.
 var Logger log.Logger = log.New()
 
-// Error implements the HTTP Problem spec laid out here:
-// https://tools.ietf.org/html/rfc7807
-type Error struct {
-	// The main error message. Should be short enough to fit in a phone's
-	// alert box. Do not end this message with a period.
-	Title string `json:"title"`
-
-	// Id of this error message ("forbidden", "invalid_parameter", etc)
-	ID string `json:"id"`
-
-	// More information about what went wrong.
-	Detail string `json:"detail,omitempty"`
-
-	// Path to the object that's in error.
-	Instance string `json:"instance,omitempty"`
-
-	// Link to more information about the error (Zendesk, API docs, etc).
-	Type string `json:"type,omitempty"`
-
-	// HTTP status code of the error.
-	Status int `json:"status,omitempty"`
-}
-
-func (e *Error) Error() string {
-	return e.Title
-}
-
-func (e *Error) String() string {
-	if e.Detail != "" {
-		return fmt.Sprintf("rest: %s. %s", e.Title, e.Detail)
-	} else {
-		return fmt.Sprintf("rest: %s", e.Title)
-	}
-}
-
-var handlerMap = make(map[int]http.Handler)
-var handlerMu sync.RWMutex
+var (
+	handlerMap = make(map[int]http.Handler)
+	handlerMu  sync.RWMutex
+)
 
 // RegisterHandler registers the given HandlerFunc to serve HTTP requests for
 // the given status code. Use CtxErr and CtxDomain to retrieve extra values set